Move progress computation to the store
